// src/app/core/api/model/game.ts
/**
 * Sogexia Hiring Test API - OpenAPI 3.0
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Score } from './score';


/**
 * One game
 */
export interface Game { 
    id?: number;
    /**
     * List of scores
     */
    scores?: Array<Score>;
}