// src/app/core/api/model/player.ts
/**
 * Sogexia Hiring Test API - OpenAPI 3.0
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * One player
 */
export interface Player { 
    id?: number;
    name?: string;
}